refactor user permission repositories
