--- src/domain/entities/user-monitoring/common/UserGroup.ts
@@ -1,4 +1,0 @@
-export interface UserGroup {
-    id: string;
-    name: string;
-}
