--- src/domain/entities/user-monitoring/permission-fixer/UserGroup.ts
@@ -1,4 +1,0 @@
-export interface UserGroup {
-    id: string;
-    name: string;
-}
